Rename Module (module-cache -> module-redis)
