+++ build.gradle
@@ -54,7 +54,7 @@ project(':module-api') {
     dependencies {
         compileOnly project(':module-core')
         compileOnly project(':module-adapter')
-        compileOnly project(':module-cache')
+        compileOnly project(':module-redis')
         compileOnly project(':module-security')
     }
 }
