+++ module-api/src/main/java/com/myfin/api/service/impl/UserSignUpServiceImpl.java
@@ -6,10 +6,10 @@ import com.myfin.api.dto.VerifyIdentity;
 import com.myfin.api.dto.VerifyIdentityResultDto;
 import com.myfin.api.service.TopServiceComponent;
 import com.myfin.api.service.UserSignUpService;
-import com.myfin.cache.entity.CacheVerified;
-import com.myfin.cache.entity.CacheVerifyCode;
-import com.myfin.cache.repository.CacheVerifiedRepository;
-import com.myfin.cache.repository.CacheVerifyCodeRepository;
+import com.myfin.redis.entity.CacheVerified;
+import com.myfin.redis.entity.CacheVerifyCode;
+import com.myfin.redis.repository.CacheVerifiedRepository;
+import com.myfin.redis.repository.CacheVerifyCodeRepository;
 import com.myfin.core.dto.UserDto;
 import com.myfin.core.entity.User;
 import com.myfin.core.exception.impl.BadRequestException;
