+++ module-api/src/test/java/com/myfin/api/service/impl/UserSignUpServiceImplUnitTest.java
@@ -4,9 +4,9 @@ import com.myfin.adapter.coolsms.SMSMessageComponent;
 import com.myfin.api.dto.SignUp;
 import com.myfin.api.dto.VerifyIdentity;
 import com.myfin.api.dto.VerifyIdentityResultDto;
-import com.myfin.cache.entity.CacheVerifyCode;
-import com.myfin.cache.repository.CacheVerifiedRepository;
-import com.myfin.cache.repository.CacheVerifyCodeRepository;
+import com.myfin.redis.entity.CacheVerifyCode;
+import com.myfin.redis.repository.CacheVerifiedRepository;
+import com.myfin.redis.repository.CacheVerifyCodeRepository;
 import com.myfin.core.dto.UserDto;
 import com.myfin.core.entity.User;
 import com.myfin.core.entity.UserAddressVO;
