+++ module-redis/src/main/java/com/myfin/redis/entity/CacheVerifyCode.java
@@ -1,4 +1,4 @@
-package com.myfin.cache.entity;
+package com.myfin.redis.entity;
 
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
@@ -10,19 +10,21 @@ import org.springframework.data.redis.core.RedisHash;
 import java.util.Objects;
 
 @Getter
-@RedisHash(value = "Verified", timeToLive = 86400L) // TTL: 1일
+@RedisHash(value = "VerifyCode", timeToLive = 180L) // TTL: 3분
 @NoArgsConstructor(access = AccessLevel.PRIVATE)
 @AllArgsConstructor(staticName = "of")
-public class CacheVerified {
+public class CacheVerifyCode {
 
     @Id
     private String phoneNum;
+
+    private String code;
 
     @Override
     public boolean equals(Object o) {
         if (this == o) return true;
         if (o == null || getClass() != o.getClass()) return false;
-        CacheVerified that = (CacheVerified) o;
+        CacheVerifyCode that = (CacheVerifyCode) o;
         return Objects.equals(phoneNum, that.phoneNum);
     }
