+++ module-redis/src/main/java/com/myfin/redis/repository/CacheVerifyCodeRepository.java
@@ -1,6 +1,6 @@
-package com.myfin.cache.repository;
+package com.myfin.redis.repository;
 
-import com.myfin.cache.entity.CacheVerifyCode;
+import com.myfin.redis.entity.CacheVerifyCode;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
