+++ settings.gradle
@@ -2,5 +2,5 @@ rootProject.name = 'my-fin'
 include 'module-core'
 include 'module-api'
 include 'module-adapter'
-include 'module-cache'
+include 'module-redis'
 include 'module-security'
